backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth>=7.1.1",
    "django==5.2",
    "django-enumchoicefield>=3.0.1",
    "djangorestframework>=3.16.1",
    "pillow==12.1.0",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
]
